Added the ability to have multiple texures in one model, multi threaded loading, and other fixes and improvements.

+++ src/Drone.hpp
@@ -82,6 +82,6 @@ private:
 	glm::vec3 _angularMomentum;
 	float _mass;
 
-	std::unordered_map<unsigned int, Engines> _engines;
+	std::unordered_map<uint64_t, Engines> _engines;
 	DronePlugin _plugin;
 };
